requirements.txt
ops
PyYAML
pytest